# docs/swagger.yaml
definitions:
  controllers.ResponseHTTP:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  user.User:
    properties:
      name:
        type: string
      password:
        type: string
      status:
        type: integer
    type: object
info:
  contact: {}
  description: This is an API for Book Application
  title: Book App
  version: "1.0"
paths:
  /api/books:
    get:
      consumes:
      - application/json
      description: Get all books
      produces:
      - application/json
      responses:
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ResponseHTTP'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/controllers.ResponseHTTP'
      summary: Get all
      tags:
      - books
  /api/books/{id}:
    get:
      consumes:
      - application/json
      description: Get book by ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ResponseHTTP'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/controllers.ResponseHTTP'
      summary: Get book by ID
      tags:
      - books
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: sign-in process
      parameters:
      - description: AAA 123
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.User'
      produces:
      - application/json
      responses:
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ResponseHTTP'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/controllers.ResponseHTTP'
      summary: sign-in
      tags:
      - auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: Sign-Up process
      parameters:
      - description: AAA 123
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.User'
      produces:
      - application/json
      responses:
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ResponseHTTP'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/controllers.ResponseHTTP'
      summary: sign-up
      tags:
      - auth
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"